+++ playground/worker/src/App.tsx
@@ -1,0 +1,10 @@
+import { useState } from "react";
+import MyWorker from "./worker-via-import?worker&inline";
+
+new MyWorker();
+
+export const App = () => {
+  const [count, setCount] = useState(0);
+
+  return <button onClick={() => setCount(count + 1)}>count is {count}</button>;
+};