+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - fix: commonjs default export (fixes [#14](https://github.com/vitejs/vite-plugin-react-swc/issues/14))
 - fix: support Vite base option (fixes [#18](https://github.com/vitejs/vite-plugin-react-swc/issues/18))
-- fix: compatibility with react-refresh runtime in @vitejs/plugin-react (#20)
+- fix: compatibility with react-refresh runtime in @vitejs/plugin-react ([#20](https://github.com/vitejs/vite-plugin-react-swc/pull/20))
 
 ## 3.0.0